xdmapcie: Fix ECAM base address size
ECAM base address can be a 64 bit address based on user selection.
Adding support for this.

Acked-by: Bharat Kumar Gogada <bharat.kumar.gogada@xilinx.com>

+++ XilinxProcessorIPLib/drivers/xdmapcie/src/xdmapcie.h
@@ -127,7 +127,7 @@ typedef  struct {
 					 */
 	u8  IncludeRootComplex;		/**< Is IP built as root complex */
 
-	u32 Ecam;
+	u64 Ecam;
 	u32	NpMemBaseAddr;		/**< non prefetchable memory base address */
 	u64	PMemBaseAddr;		/**< prefetchable memory base address */
 	u32	NpMemMaxAddr;	/**< non prefetchable memory max base address*/
